Rename to PeriodicRunnableTask, the manager can support callables in future
